Updated the font class to remove nastyness Updated the WM for new skins

+++ src/main.h
@@ -20,7 +20,7 @@ using namespace std;
 
 cGraphics		Graphics;
 
-float mousex, mousey;
+long mousex, mousey;
 float oldmousex, oldmousey;
 
 float translatedstartmousex, translatedstartmousey;
