+++ src/wateredit.cpp
@@ -2,7 +2,7 @@
 #include "graphics.h"
 #include "undo.h"
 
-extern float mousestartx, mousestarty;
+extern long mousestartx, mousestarty;
 extern double mouse3dx, mouse3dy, mouse3dz;
 extern cGraphics Graphics;
 extern cUndoStack undostack;
