+++ src/wm/windowprogressbar.h
@@ -33,8 +33,8 @@ public:
 	virtual void click();
 	virtual string ppopup();
 
-	void SetInt(int, intptr_t);
-	intptr_t GetInt(int);
+	void SetInt(int, int);
+	int GetInt(int);
 };
